@@ -1,7 +1,8 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/map'
+import { from } from 'rxjs';
+import { map } from 'rxjs/operators';
+import 'rxjs/add/operator/map';
 import { LoginRequest } from '../models/LoginRequest';
 
 @Injectable()
